Cleanup & inline migrate
