+++ packages/nx-firebase/src/generators/sync/schema.d.ts
@@ -2,5 +2,4 @@ export interface SyncGeneratorSchema {
   app?: string
   project?: string
   // functions?: boolean
-  migrate?: boolean
 }
